backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Temporal workflow engine
temporalio==1.5.1

# HTTP client for API calls
httpx==0.25.2

# Configuration and utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
